+++ EssentialFeed/EssentialFeed/Feed Cache/FeedCachePolicy.swift
@@ -1,0 +1,20 @@
+//
+//  FeedCachePolicy.swift
+//  EssentialFeed
+//
+//  Created by Cristian Felipe Patiño Rojas on 4/4/25.
+//
+
+import Foundation
+
+internal final class FeedCachePolicy {
+    private static let calendar = Calendar(identifier: .gregorian)
+    private static var maxCacheAgeInDays: Int { return 7 }
+    private init() {}
+    static func validate(_ timestamp: Date, against date: Date) -> Bool {
+        guard let maxCacheAge = calendar.date(byAdding: .day, value: maxCacheAgeInDays, to: timestamp) else {
+            return false
+        }
+        return date < maxCacheAge
+    }
+}
